Update
Cập nhật lại panel thông tin chi chuyến bay.
Cập nhật lại BLL hóa đơn nhận 1 output về thuế.
Cập nhật lại Form chọn chỗ ngồi thêm giá vé.

Sửa lại UC, thêm 1 UC mới thông tin thanh toán.
Thêm 1 class pair chứa 2 object.

+++ Models/KhachHangTreEm.cs
@@ -10,16 +10,16 @@ namespace QuanLyBanVeMayBay.Models
         private DateTime ngaysinh;
 
         private int magoihanhlychieudi;
-        private float giatiengoihanhlychieudi;
+        private double giatiengoihanhlychieudi;
 
         private int magoihanhlychieuve;
-        private float giatiengoihanhlychieuve;
+        private double giatiengoihanhlychieuve;
 
         private int mavechieudi;
-        private float giatienvechieudi;
+        private double giatienvechieudi;
 
         private int mavechieuve;
-        private float giatienvechieuve;
+        private double giatienvechieuve;
 
         public KhachHangTreEm(string hoten, string gioitinh, DateTime ngaysinh)
         {
@@ -33,12 +33,12 @@ namespace QuanLyBanVeMayBay.Models
         public string Gioitinh { get => gioitinh; set => gioitinh = value; }
         public DateTime Ngaysinh { get => ngaysinh; set => ngaysinh = value; }
         public int Magoihanhlychieudi { get => magoihanhlychieudi; set => magoihanhlychieudi = value; }
-        public float Giatiengoihanhlychieudi { get => giatiengoihanhlychieudi; set => giatiengoihanhlychieudi = value; }
+        public double Giatiengoihanhlychieudi { get => giatiengoihanhlychieudi; set => giatiengoihanhlychieudi = value; }
         public int Magoihanhlychieuve { get => magoihanhlychieuve; set => magoihanhlychieuve = value; }
-        public float Giatiengoihanhlychieuve { get => giatiengoihanhlychieuve; set => giatiengoihanhlychieuve = value; }
+        public double Giatiengoihanhlychieuve { get => giatiengoihanhlychieuve; set => giatiengoihanhlychieuve = value; }
         public int Mavechieudi { get => mavechieudi; set => mavechieudi = value; }
-        public float Giatienvechieudi { get => giatienvechieudi; set => giatienvechieudi = value; }
+        public double Giatienvechieudi { get => giatienvechieudi; set => giatienvechieudi = value; }
         public int Mavechieuve { get => mavechieuve; set => mavechieuve = value; }
-        public float Giatienvechieuve { get => giatienvechieuve; set => giatienvechieuve = value; }
+        public double Giatienvechieuve { get => giatienvechieuve; set => giatienvechieuve = value; }
     }
 }
